pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "django-db-chat"
version = "0.1.1"
authors = [
  { name="Favour", email="fagozie43@gmail.com" },
]
description = "A Django app providing a natural language interface to query databases directly via LLMs."
readme = "README.md"
license = {text = "MIT License"}
requires-python = ">=3.8"
dependencies = [
    "django",
    "anthropic",
    "psycopg2-binary",
    "redis",
]
classifiers = [
    "Environment :: Web Environment",
    "Framework :: Django",
    "Framework :: Django :: 3.2",
    "Framework :: Django :: 4.0",
    "Framework :: Django :: 5.0",
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
    "Topic :: Communications :: Chat",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

# Tell setuptools where to find the package code
[tool.setuptools.packages.find]
include = ["db_chat*"]  # Finds the db_chat package and any subpackages

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "coverage",
    "daphne>=4.0",
]
realtime = [
    "channels>=4.0",
    "channels-redis>=4.0",
]

[project.urls]
"Homepage" = "https://github.com/f-gozie/dbchatbot"
"Bug Tracker" = "https://github.com/f-gozie/dbchatbot/issues"
